requirements.txt
# LLM-AIXI Project Dependencies

# Google Cloud AI Platform
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
google-auth>=2.23.0

# Environment and configuration
python-dotenv>=1.0.0

# Web search
requests>=2.31.0

# Data handling (pathlib is part of Python standard library)

# Optional: For enhanced code execution security
# docker>=6.1.0  # Uncomment if using Docker for code execution

# Development and testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
